feat: add hook for employee validation
