+++ src/routes/holidays.tsx
@@ -1,26 +1,16 @@
-import { Route, Routes, useNavigate } from "react-router-dom";
-import { useEffect } from "react";
+import { Route, Routes } from "react-router-dom";
 
 import { HolidaysOptions } from "@pages/holidays";
 import { RequestEnjoyment } from "@pages/holidays/RequestEnjoyment";
 import { RequestPayment } from "@pages/holidays/RequestPayment";
-import { useAppContext } from "@context/AppContext";
+import { useRedirectIfNoEmployee } from "@hooks/useRedirectIfNoEmployee";
 
 import { PrivilegedRoute } from "./privilegedRoute";
 
 function HolidaysRoutes() {
-  const { selectedEmployee } = useAppContext();
-  const navigate = useNavigate();
+  const selectedEmployee = useRedirectIfNoEmployee();
 
-  useEffect(() => {
-    if (!selectedEmployee) {
-      navigate("/employees/select-employee", { replace: true });
-    }
-  }, [selectedEmployee, navigate]);
-
-  if (!selectedEmployee) {
-    return null;
-  }
+  if (!selectedEmployee) return null;
 
   return (
     <Routes>
